@@ -15,14 +15,19 @@
         }
     ],
     "require": {
-        "php": "^7.1|^8.0",
-        "guzzlehttp/guzzle": "^6.0|^7.0",
-        "illuminate/support": "^5.6|^6.0|^7.0|^8.0|^9.0|^10.0|^11.0",
-        "maennchen/zipstream-php": "^v2.0"
+        "php": "^8.1",
+        "guzzlehttp/guzzle": "^7.0",
+        "guzzlehttp/psr7": "^2.6",
+        "illuminate/support": "^10.0|^11.0",
+        "league/flysystem-path-prefixing": "^3.28",
+        "maennchen/zipstream-php": "^v3.0",
+        "psr/http-message": "^2.0",
+        "spatie/laravel-package-tools": "^1.16"
     },
     "require-dev": {
-        "orchestra/testbench": "^5.0|^6.0|^7.0|^8.0|^9.0",
-        "phpunit/phpunit": "^9.0"
+        "league/flysystem-aws-s3-v3": "^3.28",
+        "orchestra/testbench": "^8.0|^9.0",
+        "phpunit/phpunit": "^9.0|^10.0"
     },
     "autoload": {
         "psr-4": {
@@ -47,7 +52,7 @@
                 "STS\\ZipStream\\ZipStreamServiceProvider"
             ],
             "aliases": {
-                "Zip": "STS\\ZipStream\\ZipStreamFacade"
+                "Zip": "STS\\ZipStream\\Facades\\Zip"
             }
         }
     }
