@@ -5,16 +5,14 @@ namespace STS\ZipStream\Models;
 use GuzzleHttp\Psr7\Utils;
 use Psr\Http\Message\StreamInterface;
 use STS\ZipStream\Exceptions\NotWritableException;
+use STS\ZipStream\OutputStream;
 
 class HttpFile extends File
 {
     private const HEADER_CONTENT_LENGTH = 'content-length';
-    /** @var array */
-    private $headers;
 
-    /**
-     * @return int
-     */
+    private array $headers;
+
     public function calculateFilesize(): int
     {
         $headers = $this->getHeaders();
@@ -30,39 +28,24 @@ class HttpFile extends File
         return $headers[self::HEADER_CONTENT_LENGTH];
     }
 
-
-    /**
-     * @return StreamInterface
-     */
     protected function buildReadableStream(): StreamInterface
     {
         return Utils::streamFor(fopen($this->getSource(), 'r'));
     }
 
-    /**
-     * @return StreamInterface
-     * @throws NotWritableException
-     */
-    protected function buildWritableStream(): StreamInterface
+    protected function buildWritableStream(): OutputStream
     {
         throw new NotWritableException();
     }
 
-    /**
-     * @inheritdoc
-     */
     public function canPredictZipDataSize(): bool
     {
-        return (is_int($this->filesize) || array_key_exists(self::HEADER_CONTENT_LENGTH, $this->getHeaders())) &&
-            parent::canPredictZipDataSize();
+        return (is_int($this->filesize) || array_key_exists(self::HEADER_CONTENT_LENGTH, $this->getHeaders()));
     }
 
-    /**
-     * @return array
-     */
     protected function getHeaders(): array
     {
-        if (!$this->headers) {
+        if (!isset($this->headers)) {
             $this->headers = array_change_key_case(get_headers($this->getSource(), 1));
         }
 
